requirements.txt
# ===================================
# Life Assistant V4 - Dependencies
# ===================================

# Core
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram Bot
python-telegram-bot==20.7

# OpenAI
openai==1.12.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
rich==13.7.0
click==8.1.7